Cross Validator Tests: updating these tests to match and use new patched session setup functions
